feat(deploy): Experimental 'unified module' API
One import can now bring in multiple AWS services. The URL syntax
strong-arms you into using action filters, which is pretty darn
important to keep the amount of code down when you're using a bunch of
services.

This seems to work fine, but I'm marking it experimental in case I
decide to move or change the URLs.